Add plugins to the bot
Still need a way to unload classes from the class-path so that a reload
command could potentially be implemented. However, with the dangers of
using that command being fairly high, I don't personally see a reason
for one to be added. Of course plugins can implement their own form of
reload system, but I don't think there will be a core one inside
CraftBot to reload plugins.